% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dendro_data.R
\name{dendro_data}
\alias{dendro_data}
\alias{dendro_data.default}
\title{Extract cluster data from a model into a list of data frames.}
\usage{
dendro_data(model, ...)
}
\arguments{
\item{model}{object of type \code{\link[stats]{hclust}}, \code{\link[stats]{dendrogram}} or \code{\link[stats]{tree}}}

\item{...}{ignored}
}
\value{
a list of data frames that contain the data appropriate to each cluster model
}
\description{
This function provides a generic mechanism to extract relevant plotting data, typically line segments and labels, from a variety of cluster models.
}
\details{
For \code{\link[stats]{dendrogram}} and \code{\link[stats]{tree}} models, extracts line segment data and labels.
}
\seealso{
There are several implementations for specific cluster algorithms:

\itemize{
\item \code{\link{dendro_data.hclust}}
\item \code{\link{dendro_data.dendrogram}}
\item \code{\link{dendro_data.tree}}
\item \code{\link{dendro_data.rpart}}
}
To extract the data for line segments, labels or leaf labels use:
\itemize{
\item{\code{\link{segment}}}: {the line segment data}
\item{\code{\link{label}}}: {the text for each end segment}
\item{\code{\link{leaf_label}}}: {the leaf labels of a tree diagram}
}
}

