\name{rpart.branch}
\alias{rpart.branch}
\title{This is an exact copy of rpart.branch in package rpart.}
\usage{
  rpart.branch(x, y, node, branch)
}
\description{
  The original function is not exported from rpart, but we
  require it to calculate the coordinates of leaves,
  segments and labels.
}
\author{
  Brian Ripley
}
\keyword{internal}

