% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro_rpart.R
\name{rpart_labels}
\alias{rpart_labels}
\title{Extract labels data frame from rpart object for plotting using ggplot.}
\usage{
rpart_labels(x)
}
\arguments{
\item{model}{object of class "rpart", e.g. the output of rpart()}

\item{...}{ignored}
}
\value{
a list with two elements: $labels and $leaf_labels
}
\description{
This code is modified from the original plot.rpart in package rpart.
}
\seealso{
\code{\link{ggdendrogram}}

Other dendro_data methods: 
\code{\link{dendro_data.rpart}()},
\code{\link{dendro_data.tree}()},
\code{\link{dendro_data}()},
\code{\link{dendrogram_data}()}

Other rpart functions: 
\code{\link{dendro_data.rpart}()},
\code{\link{rpart_segments}()}
}
\author{
Original author Brian Ripley
}
\concept{dendro_data methods}
\concept{rpart functions}
\keyword{internal}
