% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.R
\name{rpartco}
\alias{rpartco}
\title{Compute the x-y coordinates for a tree.}
\usage{
rpartco(tree, parms)
}
\arguments{
\item{tree}{Tree model}

\item{parms}{Graphics parameters}
}
\description{
Compute the x-y coordinates for a tree.
}
\keyword{internal}
