\name{segment}
\alias{label}
\alias{leaf_label}
\alias{segment}
\title{Returns segment, lavel or leaf-label data from dendro object.}
\usage{
  segment(x)

  label(x)

  leaf_label(x)
}
\arguments{
  \item{x}{dendro object}
}
\description{
  \code{segment} returns segment data from dendro object.
}

