% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdendrogram.R
\name{theme_dendro}
\alias{theme_dendro}
\title{Creates completely blank theme in ggplot.}
\usage{
theme_dendro()
}
\description{
Sets most of the \code{ggplot} options to blank, by returning blank \code{theme} elements for the panel grid, panel background, axis title, axis text, axis line and axis ticks.
}
