\name{tree.depth}
\alias{tree.depth}
\title{This is an exact copy of tree.depth in package rpart.}
\usage{
  tree.depth(nodes)
}
\description{
  The original function is not exported from rpart, but we
  require it to calculate the coordinates of leaves,
  segments and labels.
}
\author{
  Brian Ripley
}
\keyword{internal}

