% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro_tree.R
\name{tree_segments}
\alias{tree_segments}
\title{Extract data frame from tree object for plotting using ggplot.}
\usage{
tree_segments(model, uniform, ...)
}
\arguments{
\item{model}{object of class "tree", e.g. the output of tree()}

\item{...}{ignored}
}
\description{
Extract data frame from tree object for plotting using ggplot.
}
\seealso{
\code{\link[=ggdendrogram]{ggdendrogram()}}

Other tree functions: 
\code{\link{dendro_data.tree}()},
\code{\link{get_data_tree_leaf_labels}()},
\code{\link{tree_labels}()}
}
\author{
Code modified from original by Brian Ripley
}
\concept{tree functions}
\keyword{internal}
