% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro_tree.R
\name{treeco}
\alias{treeco}
\title{Function copied from tree:::treeco.}
\usage{
treeco(tree, uniform)
}
\arguments{
\item{tree}{tree object}

\item{uniform}{???}
}
\description{
Function copied from tree:::treeco.
}
\keyword{internal}

