\name{get_data_tree_leaf_labels}
\alias{get_data_tree_leaf_labels}
\title{Extract labels data frame from tree object for plotting using ggplot.}
\usage{
  get_data_tree_leaf_labels(model, ...)
}
\arguments{
  \item{model}{object of class "tree", e.g. the output of
  tree()}

  \item{...}{ignored}
}
\description{
  Extract labels data frame from tree object for plotting
  using ggplot
}
\seealso{
  \code{\link{dendro_data}}
}
\keyword{internal}

