% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro_data.R
\name{is.dendro}
\alias{is.dendro}
\title{Tests whether an object is of class dendro.}
\usage{
is.dendro(x)
}
\arguments{
\item{x}{Object to check}
}
\description{
Is a dendro?  Tests whether an object is of class dendro.
}
\seealso{
\code{\link[=dendro_data]{dendro_data()}} and \code{\link[=ggdendro-package]{ggdendro-package()}}
}
