% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro_rpart.R
\name{rpart_segments}
\alias{rpart_segments}
\title{Extract data frame from rpart object for plotting using ggplot.}
\usage{
rpart_segments(x, ...)
}
\arguments{
\item{...}{ignored}

\item{model}{object of class "tree", e.g. the output of tree()}
}
\description{
Extract data frame from rpart object for plotting using ggplot.
}
\seealso{
\code{\link[=ggdendrogram]{ggdendrogram()}}

Other rpart functions: 
\code{\link{dendro_data.rpart}()},
\code{\link{rpart_labels}()}
}
\concept{rpart functions}
\keyword{internal}
