\name{rpart_segments}
\alias{rpart_segments}
\title{Extract data frame from rpart object for plotting using ggplot.}
\usage{
  rpart_segments(model, ...)
}
\arguments{
  \item{model}{object of class "tree", e.g. the output of
  tree()}

  \item{...}{ignored}
}
\description{
  Extract data frame from rpart object for plotting using
  ggplot.
}
\seealso{
  \code{\link{ggdendrogram}}

  Other rpart functions: \code{\link{dendro_data.rpart}},
  \code{\link{rpart_labels}}
}
\keyword{internal}

