\name{rpartco}
\alias{rpartco}
\title{This is an exact copy of rpartco in package rpart.}
\usage{
  rpartco(tree,
    parms = paste(".rpart.parms", dev.cur(), sep = "."))
}
\description{
  The original function is not exported from rpart, but we
  require it to calculate the coordinates of leaves,
  segments and labels.
}
\author{
  Brian Ripley
}
\keyword{internal}

