% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro_data.R
\name{segment}
\alias{segment}
\alias{label}
\alias{leaf_label}
\title{Returns segment, label or leaf-label data from dendro object.}
\usage{
segment(x)

label(x)

leaf_label(x)
}
\arguments{
\item{x}{dendro object}
}
\description{
\code{segment} extracts line segments, \code{label} extracts labels, and \code{leaf_label} extracts leaf labels from a dendro object.
}
\seealso{
\code{\link{dendro_data}}
}
