\name{tree_segments}
\alias{tree_segments}
\title{Extract data frame from tree object for plotting using ggplot.}
\usage{
  tree_segments(model, ...)
}
\arguments{
  \item{model}{object of class "tree", e.g. the output of
  tree()}

  \item{...}{ignored}
}
\description{
  Extract data frame from tree object for plotting using
  ggplot
}
\seealso{
  \code{\link{dendro_data.tree}}
}
\keyword{internal}

